requirements.txt
openai>=1.30.0
pyyaml>=6.0
anthropic>=0.30.0
